Move metadata into src/metadata
CRA annoyingly refuses to import anything from outside that folder.
